issues

+++ 3.Implementation/src/subtract.c
@@ -1,8 +1,8 @@
 #include "header.h"
 
- int subtract(int num1,int num2)
+float subtract(float num1,float num2)
 {
-	int sub=0;
-	sub = num1-num2;
+	float sub=0;
+	sub = num1 - num2;
 	return sub;
 }
